backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
Faker==13.12.1
flake8==4.0.1
importlib-metadata==4.11.4
Markdown==3.3.7
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
zipp==3.8.0
